Se coloco servicios y caja chica

+++ back/database/migrations/2024_08_29_101918_create_productos_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('marca_pro')->nullable();
             $table->string('nombre_pro')->nullable();
             $table->string('descripcion_pro')->nullable();
-            $table->integer('precio_pro')->nullable();
+            $table->decimal('precio_pro', 10, 2)->nullable(); 
             $table->text('foto_pro')->nullable();
             $table->string('estado_pro')->nullable()->default('ACTIVO');
             $table->unsignedBigInteger('user_id')->nullable();
